Implement AbortController to reduce flakiness
Currently, the doUpdate implementation is throttled.
That means that even the latest forced updates won't fire
while the previous update is running.

This leads to a situation that the could be a stale update
running that gets intercepted by the TestRunner using a sniffer
that thinks it's the new update. Thus, the test might
proceed to the next step before the actual update is done.

This CL adds an AbortController to allow cancelling the stale
updates. It helps to avoid redundant work but also improves
testability.

Depends on https://chromium-review.googlesource.com/c/chromium/src/+/3752572

Bug: 1343290
Change-Id: I7740e3c2a97983060514770ee04e53b48fb7cbb9