Tweak DevTools display of TLS key exchange information
This removes "key exchange group" from the DevTools UI, which is kind of
a weird terminology.

For background, A TLS connection negotiates a cipher suite and, when
doing an ephemeral ECDH key exchange, a "named group". In TLS 1.2, the
cipher suite is named like TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256. The
DevTools protocol tried to decompose this name and calls the "ECDHE_RSA"
portion the "keyExchange", because it determined the rough shape of the
key exchange portion of the handshake. (A keyExchange of "RSA" meant a
very different handshake set.)

But ECDHE_RSA was still parameterized by a named group (e.g. X25519).
The name "keyExchange" was already established, so this ended up being
"keyExchangeGroup". ("Group" here refers to the abstract algebra notion
of group. Even that was probably mis-named as we really read these as
general KEM instances.)

Then, starting TLS 1.3, the cipher suites are named like
TLS_AES_128_GCM_SHA256. The handshake shape is now implicit in the
protocol.  keyExchange is empty and we only have keyExchangeGroup.

We made the DevTools UI just match this representation, with an optional
"Key exchange" and "Key exchange group" rows. But "key exchange group"
isn't common terminology and, in TLS 1.3, something like "X25519" is
better labelled as "key exchange" than "key exchange group" anyway.

Combine the two fields when displaying in the UI. If we have just a
keyExchange (the legacy "RSA" key exchange), this CL keeps things as-is.
If we have just keyExchangeGroup (TLS 1.3), just call that "Key
exchange" in the UI and drop the confusing "group" word. If we have both
(TLS 1.2 ECDHE), match the handling of the cipher row and stick the word
"with" in between.

Screenshots:
https://drive.google.com/drive/folders/1L_rR932KA4ZdLwQvbEI9y1Tiv1RWohv1?resourcekey=0-7CTlJXp3vuz1zJnHUcXa6w&usp=sharing

Bug: none
Change-Id: I433799124ea9fff8be63fe3d812fc066f65e87c5