Parse `x_google_ignoreList` in source maps
Chrome DevTools is introducing a source map extension to identify
third-party sources in source maps. The `x_google_ignoreList` field
refers to the `sources` array, and lists the indices of all the known
third-party sources in that source map.

This metadata is intended to be used for automatic ignore-listing.

The Angular CLI will introduce support for `x_google_ignoreList` in:
https://github.com/angular/angular-cli/pull/23545

Bug: 1323199
Change-Id: I0a340be4742a17b0d2b96107e650dd4421d59649
Signed-off-by: Victor Porof <victorporof@chromium.org>