Import http/tests/devtools/security tests from Chromium
Ideally these would be converted to one of the lighter test frameworks,
as they're just unit testing the UI logic against expected HTML, but
start with a simple import for now, so we can at least avoid the need
for three-sided CLs.

Modified slightly from the original versions to satisfy DevTools
presubmits.

Bug: none
Change-Id: I4c0e0917a9c9e3f4015c3ad50d31fa8e49f8bfad