Persist highlight when Memory Viewer resized
Previously, if the user enlarged the memory viewer enough, it would
try to render bytes outside of the retrieved range, which would trigger
the `memoryrequest` event. This would update the data passed to the
inspector without passing the `highlightInfo` data, which would
effectively remove the highlight. This CL makes it so that
`highlightInfo` is passed to the inspector. This makes the
highlight persist regardless of resizing.

Bug: 1336568
Change-Id: I30aea88625432870ed2cf3a4af22c36b7268d4c6