Set scroll bar behaviour for legacy data grid
This hides the horizontal scroll bar for legacy data grids. In the
standard case with automatically resizing columns no scroll bars are
needed. In the case of fixed column widths, instead of showing a
horizontal scroll bar (and potentially very narrow columns for the
non-fixed width columns), set a minimum width instead.

Fixed: 1330482, 1318375
Change-Id: I42603768887ec6e3206c525824a7cd0a2c0304f7