Persist memory highlight on debugger paused event
Previously, we would remove any highlight after a debugger paused event
as we did not know if the highlighted memory object is still in scope.
This CL implements highlight persistance by making a call to V8 after
every debugger paused event to evaluate the highlighted variable, if
one exists.

In certain cases, we cannot guarantee if the object returned by V8
is the same as the currently highlighted one. We check the type as
a basic pre-caution. However, if we track a variable in an inner
scope and exit to the parent scope, with a separate, but identically
named and typed variable, the inspector will switch the highlight
to this variable. Future work will introduce UX changes to make
this clearer to the user.

Bug: 1336568
Change-Id: I610f3c36c48778974874ae800b481cb8786f45fc