[Delegate-CH] (2) Add tests to devtools
This CL verifies that the issues sent by the accept-ch tag are still
dispatched by the delegate-ch tag. No functionality change should be
required for this to work in devtools.

This CL is part of a series:
(1) Disable support in chromium
(2) Remove tests from devtools
(3) Remove support from chromium

Commit: false
Bug: 1334152
Change-Id: I2e9948e97a83f2a28fa8530be514e260a355b7bc