Add chromeRole property to AXNode devtools protocol.
AxNode currently passes a simplified version of node roles to make them
more readable for the developers. This CL adds |chromeRole| which will
pass the raw Chrome role of each node. The immidiate use case is for AI
models that analyze the tree structure and extract the main content.
See more in http://go/chrome-screen-ai.

The Chromium side of this update is done in crrev.com/c/3771943

Bug: 1278249
Change-Id: I7063383a5f483f17bf07b9e3a5ef64d20610f2a5