Fully hide the default SplitWidget resizer element
The resizer element created by SplitWidget when it is created is not
hidden when hideDefault resizer is called. The method name suggests that
the resizer should be hidden, but it is not which can lead to issues due
to the element's high z value stealing focus with elements that cross
the divider.

Removing the element completely could also be a reasonable solution,
but since this seem to be the designed behavior and this is a legacy UI
component there seems to be a reasonable chance of a user of this
class continuing to use the default resizer after it has been 'hidden'.

Bug: 1346065
Change-Id: Ic786cbde176ca6c6f58f4c3750305e5c4fcc3176