[debugger] Type sourceMapURL as string
The sourceMapURL reported by the "script parsed" event is not
necessarily a valid Url as V8 just forwards whatever users put in the
sourceMappingUrl comment.

This CL changes the type back into a string. The actual conversion
into a Url happens in `SourceMapManager#resolveRelativeURLs` that
also does validation using `ParsedUrl`.

R=kimanh@chromium.org

Bug: 1342722
Change-Id: I4aad2865a000da2a3c89bc1b730d68a662bd8498