Move highlightInfo caching from View to Controller
This CL changes where highlightInfo is stored and
how it gets passed from the Controller to the UI components.
Previously, highlightInfo was cached inside of
the LinearMemoryInspectorView instances. This required
a clunky data-flow between the Controller and the
InspectorView. Instead, this CL moves the storage
of highlightInfo to the Controller and exposes a getter
function for the UI components to use.

We store highlightInfos in a Map
with bufferIds as keys. This will be easy to extend to
support storing multiple highlights. To make access
possible from the InspectorView, we pass it its bufferId.

Bug: 1336568
Change-Id: Ib6387a174701e9a720b5602100f0e7d5454f1ac5