Prevent host bindings from loading resources from UNC paths
This CL prevents loading of source maps and other resources
from Windows Shares via file:// UNC URLs.

We show a warning similar to other failed source map loads, for
example when the source map was not found (404).

Screenshot: https://imgur.com/a/f55Ki9A

The CL also adds a setting that allows the user to disable this
behavior. Please note that the 'host' module can't have settings
themselves, otherwise we'd have a circular dependency between 'host'
and 'common'. That is why the CL places the setting in the 'network'
module, as that is also the bottle-neck for the `loadNetworkResource`
host binding. Even though the setting is in the 'network' module, the
CL places it in the sources category, as source map loading is the
dominant use case.

R=bmeurer@chromium.org, jarin@chromium.org

Fixed: 1342722
Change-Id: I878e77592703cbfa97828650f256c2b4e4a2ac0d