Fix regression: can't expand `window` object anymore
When creating the DOM pinned properties dataset, certain types are
omitted during the `minimize` process before writing to the JSON file,
because they don't contain any information (no properties, etc.). When
checking for the mixin-included properties, this wasn't accounted for.
This caused a type error to be thrown, which this CL fixes.

Bug: 1325812
Change-Id: I937f6651b137ad19e800d736289f6276e97549d1
Signed-off-by: Victor Porof <victorporof@chromium.org>