Fix CSSProperty.formatStyle for declarations that contain a CSS function
* Changes tokenType checks to be strict equal (AFAIU from the tokenizer code[1]
we can be strict about the token types checked -- there aren't such
things like "meta-etc.")

The reason this was happening is we were checking whether
the current token is a property start using `includes` for
`variableName`. However, CSS functions like `var()` or `rgb()`
were represented with tokenType `variableName.function` thus
when we see a CSS function, we were marking current token as
a property start which is not true for that case.

[1]: https://github.com/codemirror/legacy-modes/blob/main/mode/css.js

Fixed: 1334564
Change-Id: I0eb190f6f56cfac7886640f81c60657d4d055da3