Persist memory highlight on debugger paused event
Previously, we would remove any highlight on a debugger paused event
as we could not tell if the highlighted memory object is still in
scope.
This CL implements highlight persistence by making a call to V8 after
every debugger paused event to evaluate the highlighted variable, if
one exists.

In certain cases, we cannot guarantee if the object returned by V8
is the same as the currently highlighted one. We check the type as
a basic pre-caution. However, if we highlight a variable X in scope
A and resume the debugger until it stops in scope B with variable Y
where X and Y have the same name and type, the inspector will
switch the highlight from X to Y.

Bug: 1336568
Change-Id: I610f3c36c48778974874ae800b481cb8786f45fc