DataGrid: make scroll sync with render
Currently, automatic scroll to bottom is broken (in ProtocolMonitor) if many events arrive at
the same time. This happens because scrolling happens asynchroniously
and the component might detect that the user is not at the bottom of the page
because the scroll is still being scheduled while another scroll caused
by new data is happening. It looks like removing coordinator from the loop
does not actually degrade the performance and also fixes the automatic scroll.

Bug: none
Change-Id: I5c4e2b9872d6bded45a85e5363ff7404cb82b186