DataGrid: make scroll sync with render
Currently, automatic scroll to bottom is broken (in ProtocolMonitor) if many events arrive at
the same time due to the fact that scrolling happens asynchronously
and the component might detect that the user is not at the bottom of the page
because the scroll is still being scheduled while another scroll caused
by the arriving events is happening. It looks like removing coordinator from the loop
does not actually degrade the performance and also fixes the automatic scroll.

Bug: none
Change-Id: I5c4e2b9872d6bded45a85e5363ff7404cb82b186