Add Interactions to Performance Panel
This CL adds an "Interactions" track that will show interactions. There
is currently no additional highlighting for interactions that are deemed
"long", or to show a website's INP score.

An interaction event will show its ID (useful for linking multiple
events together - e.g. a `pointerdown` and `pointerup` that were the
same user interaction) and the type of event (`pointerdown`).

By rendering these onto the timeline, the user can visually align them
with the main thread tasks, and use that to track down what may be the
issue:

https://imgur.com/a/6epEJUf

Bug: 1347390
Change-Id: Ic38797434172e5e85e3cd993fcd2ca25a9f565ea