[LMI] Add HighlightChipList component
This CL adds a list of chips to the memory inspector which
display information on the inspected expressions when
debugging with the C/C++ DWARF Extension. A chip lets the
user to jump to the start of the expresison's highlighted memory and
to delete the highlight.

Before/After: https://imgur.com/a/yticXm6
Design Doc: https://bit.ly/3PM7mF1
Bug: 1336568
go/chrome-devtools:extending-memory-inspector-design
Change-Id: Ifbd461ead98b6461d1e3a4406a9ca161425aceac