Deflake extensions test
Make sure the resource has actually been loaded before trying to open
it through the extension API.

This CL adds a separate utility to wait for source files loaded or added. The advantage of the new version is that it is nestable, i.e., it can be used within helpers without the test having to be aware of this.

Fixed: 1350929
Change-Id: Ifd76761d50a58cacb6cbcf9f7bf51ebde51ca546