Fix Interactions detection in Performance Panel
It seems that we can have many processes with the name `Renderer`, and
so when we look up based on name, we don't always get the right process.
I think this is a broader bug within the Perf Panel, but I don't want to
fix that immediately because I'm worried about breaking code that relies
on this slightly incorrect logic (as far as I can work out, when you
look up by name, you just get whatever `Renderer` process was last named
in the trace).

Therefore, to fix just interactions, the code instead loops through all
processes and then the `CrRendererMain` thread within each of them, to
find the interactions.

Bug: 1347390
Change-Id: If2222b87e87fdcda6a2634fe1607768e530b4387