[LMI] Prepend asterisk to pointer variables
When a user inspects a variable in the LMI, the Inspector
displays the pointed-to object. This CL prepends an asterisk
in front of pointer inspected pointer variables to indicate
that we display the pointed-to object. This feature was inspired
by C++ dereferencing.

Before/After: https://imgur.com/a/JI4jLI4
Design Doc: https://bit.ly/3PM7mF1
Bug: 1336568
go/chrome-devtools:extending-memory-inspector-design
Change-Id: Icc2131a9ede5b4b208a55ee0289c2a4309a428a3