Deflake raw wasm e2e tests
The "stepping with state" tests time out because sometimes breakpoints aren't removed. This is likely caused by the "scroll to line" step that was part of breakpoint removal. Scrolling was implemented by bringing up the Go To Line dialog with a hotkey. It seems that filling out and closing the dialog sometimes doesn't work, which then causes the click that was supposed to remove the breakpoint to only close the dialog. Hence, remove scroll to line, since it's not required in any of the tests anyways.

Drive-by: unskip the raw-wasm test, since the flake didn't reproduce in hundreds of iterations.

Fixed: 1349290, 1349297, 1321898
Change-Id: I1b75d20ae99dea11d1929e11157da0253f4736cc