Enable silent replMode for live expressions
We disabled REPL mode in the past due to interacting with
"Pause on exception". What I missed is that Runtime.evaluate already
has a parameter "silent" for this use case: the inspector will ignore
exceptions and promise rejections triggered by the evaluated expression
and don't pause even with "Pause on exception" generally enabled.

The V8 CL https://crrev.com/c/3822685 adds a CDP level test for this
behavior.

R=bmeurer@chromium.org

Fixed: 1335439
Change-Id: I198806576d95621f523020ce512974b8b895f370