[LMI] Show full name path when inspecting via context menu
Previously, when inspecting properties of an object through the
context menu, only the name of the property would be passed to
the LMI. For example, when inspecting `myObject.subObject` via
the context menu, only `subObject` would be passed to the LMI.
This CL fixes this bug so that the full name path
`myObject.subObject` is passed along.

Drive-by: Removed an unused parameter from `openInspectorView()`
and renamed related arguments and parameters for consistency.

Before/After: https://imgur.com/a/iu9iLzG
Design Doc: https://bit.ly/3PM7mF1
Bug: 1336568
go/chrome-devtools:extending-memory-inspector-design
Change-Id: Ia4478f1f6099c39f36b7cced089f1f42cfc4434e