Reduced redundant if conditions
The 1st condition in upsized if condition is always true if 2nd is true.
You might thing that the firs condition is enought. With the 1st
we know that the 2nd would always be true till the rest of the function.
IDE doesnt know that, and if I want to ommit redundant if conditions it
shows an error that #this.EmulationModel might be null.
With this change we have only one redundant if condition and not two.


Bug: none
Change-Id: I6f756f613628c1f35e7762e8e5364d9dfa0726c8