Reduced redundant if conditions
The 1st condition in up-sized if-condition is always true if 2nd is true. You might think that the 1st condition is enough, but that way you would need to check for the 2nd condition at least twice in the code.
With the 1st condition we know that the 2nd one would always be true till the rest of the function. IDE doesn't know that, and if I want to omit redundant if conditions later in the function, IDE shows an error that #this.EmulationModel might be null.
With this change we have only one redundant if condition and not two.


Bug: none
Change-Id: I6f756f613628c1f35e7762e8e5364d9dfa0726c8