fullscreen screenshots are taken with a flag beyondviewport = false
this change makes fullscreen screenshot be taken with the CDP
if we want to take fullpage screenshots without passing a clip.

Without the change, fullscreen and fullpage screenshots gave the
same picture, and in case of fullscreen the pictured was cropped.
The cropping wasnt always the best.

After the change, fullscreen screenshots get the picture already
cropped by cdp, because it uses within viewport and not beyond.

Bug: none
Change-Id: I0ab13a3e15c0eb79aa37509f3166a12f4d085d8f