Add info about non-overridden response headers to `NetworkRequest`
We want to be able to distinguish between overridden and non-overridden
headers in the network panel's `RequestHeadersView`. This CL adds a new
property `nonOverriddenResponseHeaders`(naming TBD) to `NetworkRequest`,
which is populated when fulfilling an intercepted request (potentially
with overridden headers), so that the original/non-overridden header
values can still be accessed.

Bug: 1297533
Change-Id: I613f98398633269313cf7a5bcfa9dda917f6f6d9