Fix top layer not loading issue
Currently if we don't have an ElementsTreeElement inserted in the tree,
we would not be able to add its top-layer adorner, which short-circuits
the rest of the logic of adding the node to the list of top layer nodes.
This CL fixes this logic, but would need a separate CL to delay the
adornment of these ElementsTreeElements when they are eventually
created (a delayed adornment).

Fixed: 1352407
Change-Id: Ifc5ec5b58cd6cc319cd78c7ce64fb134482fd829