Preparation for change of behaviour
While taking a screenshot of viewport we call CDP with
beyondviewport = false because if we were to call with no clip
and beyondviewport = true, after the change of the behaviour in CDP, we would have a fullpage screenshot.

Viewport screenshot goes through the part that is commented to be deleted, to show that it would work with CDP change.
We cannot create a final change of behaviour right now because we are still waiting for CDP change.

Design Doc: https://docs.google.com/document/d/11Dwvy3UuGaFdBxE-nWzVG-oDSQTvEyomu42SN3dCpRs/edit#heading=h.8btw5sz9igs9
Related CL(s): https://chromium-review.googlesource.com/c/chromium/src/+/3797208

Bug: chromium:1357584
Change-Id: I01f78ac0269b7428c923e1292817e8a58bac3242