Fix encoding when renaming files
This allows snippets with special characters in their name to be
deleted.

This is a follow up to https://crrev.com/c/3613869. Previously, when
creating or renaming a file, the UI tree showing the file name would
internally store the file name with a different encoding. When trying
to delete the file/snippet, the file would be deleted on disk, but
the file name would not be found among the tree nodes, causing the
tree node to not be deleted. This is now fixed.

Fixed: 1353321
Change-Id: Ie160b2ec6840be54a05bb8111cbe2d145a6d9338