Fix keyboard focus order for Application tab Manifest section
In the Manifest section of Application tab, the keyboard focus for its
sub list item is not set to the appropriate section which has been
scrolled in to view. Screen reader also will not benefit from the
scroll of the viewport.
Action:After selecting a Manifest submenu item and tabbing to the right
Previous:The focus goes to the top of Manifest causing the view to
scroll back all the way to the top of Manifest.
After the fix: The focus goes to that selected sub-section.

Bug: 1353891
Change-Id: I8f11bdb19198b6bf303552257f50e675d55ed115