[BreakpointSidebarPane] Show re-designed sidebar
This CL makes use of the BreakpointsView in order
to show breakpoints if the experiment is enabled, see:
'Enable re-design Breakpoint Sidebar Pane in the Sources
Panel'

How to reproduce:
* Enable experiment
* Navigate to Sources Panel
* Set a breakpoint in a source file

Limitations:
Currently all interaction within the panel are not
forwarded to the internal logic. This CL is only
for showing the new pane.

Drive-by: CSS padding fix
Before/After: https://imgur.com/a/mKKOvYs
Doc: https://bit.ly/3pa4y92
Bug: 1346231, 1324904
Change-Id: Ia209b6d395e79a47696c659ad14eff3076cf5541