Fix Simplified Chinese translation for i18n
When I tried to replicate a request in the Network tab of the Chrome console, I noticed a little problem with its menu content:
https://s1.ax1x.com/2022/08/31/v4hN4O.png
Node.js and fetch are two words that should have a space between them, like the fourth one below it.
So I fixed it, and checked the Hong Kong and Taiwan translations, and they were correct, with a space in between.

BUG=nobug

Change-Id: Ie7788530ef0b45ded431fd353450f64b095202ab