[docs] Reorganize project documentation.
Going forward we want to have DevTools project documentation maintained
within the source repository, following the model that Chromium uses for
its project documentation[1].

This change sets up the basics for that, putting in place the main
entry point (in docs/README.md) and moving the architecture
documentation to the docs folders, as well as putting back the 3rd party
guidelines here. We also adopt the simpler structure for images, making
it easier to identify which document a given image file belongs to.

[1]: https://chromium.googlesource.com/chromium/src/+/main/docs/README.md

DISABLE_THIRD_PARTY_CHECK=reorganizing files

Fixed: chromium:1358806
Change-Id: I70f7f90b95c7311ca24bede04ca9cdd5e4c31eaf