Fix source map loading for Node.js
Loading source maps requires checking the value of a setting. This
setting was put into the Network module since its related to loading
source maps over the network.

This breaks Node.js debugging since the Node.js DevTools doesn't have
a network panel so the setting gets never registered. This causes all
source map loads to fail.

The solution is to put the setting into `sdk` which is guaranteed to
be loaded everywhere.

R=kimanh@chromium.org

Fixed: 1358497
Change-Id: I554d2fe7774a620d7cf5b19d3f9bcae5bfd246e5