[LMI] Prototype for multiple highlights
This CL is WIP. It implements a prototype for multiple highlighting.

Some of the work included in this CL:
* Caching lists of HighlightInfo
* Passing around lists of HighlightInfo
* Modifying tests to work with multiple highlights
* Changing how logic for chip focusing works

This exploration discovered some pitfalls with multiple highlights:
For instance, previously, clicking a chip would make the Viewer
jump to the object's first byte. This would in turn be used to
select the smallest enclosing highlight. However, with multiple
highlights, it can happen that two pointers point to the same object.
It would be silly if clicking a highlight chip of object A would
result in chip B getting focused. I implemented a prototype where we
instead pass around the HighlightInfo object and decide the highlight
based on it.

Clicking on a byte that belongs to multiple overlapping objects
creates an issue: it isn't clear which object should be focused.
It could be the earliest inspected one, the latest one, the smallest
enclosing one, or potentially all of them. Focusing all of them would
probably be the safest bet, but this would then probably require to
add a way to manually de-focus an object through some toggle button.

Change-Id: Ib815457ffe6831b11c0f95383700a17566a008b3