[SourceMap] Trim away BOM at the beginning of the sourcemap file
When loading the source maps we may get a file that has a BOM note at
the start. Passing the loaded content to the JSON.parse functionality
will then fail, as it doesn't expect the BOM at the start.

This edge case only happens in some cases: only if we make use of the
MultitargetNetworkManager, which uses the NetworkResourceLoader in the
back-end, which passes the read content/bytes as is to the
devtools-frontend.

The CL adds a step before parsing the JSON and trims away the BOM
note if existent. We decided to strip it in the front-end since
we know that in this particular case we are dealing with json and
this doesn't allow a BOM at the beginning. Stripping it in the back-end
would strip it for all use cases.

Also-by: mathias@chromium.org
Fixed: 1353983
Change-Id: I466ef9a07084a4cc4dcfec8e650e55e01ce6216a