Make elements tree ellipsis clickable to expand
The ellipsis in the elements tree panel isn't clickable now.
Giving a clear button for expanding is more in line with human intuition. And with the button in the middle, it's easier to click than the left triangle.

Screenshot: https://i.imgur.com/krzohwA.jpg
Screen Recording: https://i.imgur.com/XejrhXW.mp4

Bug: None
Change-Id: I5fbd78dff38490af08fdb01c5cd8a1ed063f3974