[BreakpointsSidebarPane] Add listeners for interacting with sidebar
This adds listeners to the BreakpointsSidebarPane for jump to source
and enabling/disabling the checkbox, such that interacting with the
sidebar actually propagates the action to the SourcesPanel (jump)/
BreakpointManager (enable/disable).

This is still behind an experiment.

Screenshot: https://imgur.com/a/CT0JqLA
Doc: https://bit.ly/3pa4y92
Bug: 1324904, 1346231
Change-Id: Ic620d1c24cef465c81ffd77fb06a425752971555