[Performance] Translate function names in the timeline flame chart
The implementation is analogous to that of [0]. However, since the
source of the call frame data is different, (Debugger domain vs Sampling
data) extra steps have to be taken to extract the script's information.

In particular the 'target' or the context of the sample has to be
provided, once the target is availble, the corresponding script and
source map can be queried.

[0] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/3720713

Bug: 1364601
Change-Id: I334db51bad622ce4bc81d95f8da869e383811cf3