Allow modifying headers in the Network Panel
This CL allows the user to edit existing headers in the network
panel's headers view. The edits will be added to the header overrides,
such that upon refreshing the page, the response headers contain the
edited values.

Screencast: https://i.imgur.com/zlvzRyE.mp4

The edits are stored to disk and sent to the backend when the editable
element loses focus.

If there are multiple headers with the same name e.g. 'foo', these are
treated as a single unit. This means that when overriding them, all
'foo' headers are removed and replaced with the override(s) for 'foo'.

Bug: 1363132
Change-Id: Ib2ee738bfbe0974d12f0f854210afaa16f0a7259