Autocomplete CSS property name on pressing space
Currently, adding space in property names
gives us an undefined behavior and can be
a cause of funny bugs (See Fixed)

This CL changes the behavior of CSSPropertyPrompt
to accept autocomplete suggestion on space
press just like tab press.

The reason of the bug was, `TextPrompt`
has space as the stopping character. Thus,
when it was providing autocompletions for
`a 1` it was only providing auto completions
for value `1` and replacing `a 1` with
`a font-weight: 100` which in turn
adds style `a\nfont-weight: 100` to
the stylesheet which causes an unwanted
behavior described in the bug

Fixed: 1343316
Change-Id: I16f260dcb32f65859783bd4b72e3cd3b7b0f267a