Fix jumping to a wrong location after tab change
Before this, what was happening is the following:
* We were setting `scrollDOM` scroll properties
and it was scrolling directly
* After another update in CodeMirror, it was
calling `requestMeasure` which, while making
the measurement, also updating the scrollDOM
scroll position for some reason.

This CL fixes that situation by using
`scrollIntoView` effect which doesn't
cause `CodeMirror` to update the
`scrollDOM` scroll position.

See the fixed behavior in here[1]

[1]: https://bugs.chromium.org/p/chromium/issues/detail?id=1368113#c1

Fixed: 1368113
Change-Id: Ib5e6e0272907d7909043ed0226e4cc3c52e8e0cb