Fix network log autoscroll regression
The recent changes to improve accessibility for the legacy DataGrid
broke the custom scroller for the network log DataGrid. It couldn't be
scrolled all the way to the bottom. Aside from the visual regression,
this also broke the network log's autoscroll behavior.

This applies the recommended fix from http://crbug.com/1349531
The waterfall scroller's height needs to exclude the header height
since this should not be counted as a part of the scrollable content.

Bug: 1349531
Change-Id: I09a022dc38c14b5b3b3910e4e5fb43a584473c47