Accessibility fix - Avoid marking separators as menu items
Bug: 1368709

The soft context menu was marking everything underneath itself as a
menu item checkbox when it is created, including any menu section
separators it contains. Screen readers didn't read out the empty string
separator values, but they did affect the count.

This lead to potential confusion for users as the reader skipped from
item 3/10 to item 5/10 with no way to access or see what happened to
item 4/10. In this same situation, this will now correctly go from 3/9
to 4/9 since the separator is no longer counted.

Change-Id: I22d3fe59f26412f3879d5cc2121c46da57269ed7