Show why Chrome uses a specific transport protocol for HTTP semantics in the DevTools.
Note that this CL is transferred to https://crrev.com/c/3924610.

In some cases, Chrome fails to use HTTP/3 even when the server
supports HTTP/3, so it'll be useful to show the reason in the DevTools.

This new feature is valid only when the experimental flag "ALTERNATE_PROTOCOL_USAGE" is on.

Design Docs is here:
https://docs.google.com/document/d/1SfMy7MMTbSn2C1RkPxnlnQdj9AnZ5-fQdEl0XO1s2ZQ/edit?usp=sharing

DISABLE_THIRD_PARTY_CHECK=tentative

Bug:1345536
Change-Id: I98483fbfee713a2e6193faba237f4332e9192394