Show why Chrome uses a specific transport protocol for HTTP semantics in the DevTools.
Note that this CL is transferred from https://crrev.com/c/3889955.

In some cases, Chrome fails to use HTTP/3 even when the server
supports HTTP/3, so it'll be useful to show the reason in the DevTools.

This new feature is valid only when the experimental flag "ALTERNATE_PROTOCOL_USAGE" is on.

Design Docs is here:
https://docs.google.com/document/d/1SfMy7MMTbSn2C1RkPxnlnQdj9AnZ5-fQdEl0XO1s2ZQ/edit?usp=sharing

DISABLE_THIRD_PARTY_CHECK=tentative

Bug:1345536
Change-Id: I5d1e03a12762812dce8b281b60d6697f9a215015