[BreakpointsSidebarPane] Add pause on (caught) exceptions buttons.
This adds the "Pause on exceptions" and "Pause on caught exceptions"
checkboxes to the new breakpoints sidebar pane, keeping the existing
semantics, and thus aligning it to the UX found in Firefox DevTools.

This change also disables the previous functionality in the Sources
panel when the experiment is turned on.

Drive-by-fix: Refactor the BreakpointsSidebarController a bit to make
the tests a bit easier to understand.

Fixed: chromium:1324920
Bug: chromium:1266395,chromium:1346231
Screenshot: https://imgur.com/DGQ1rtG.png
Doc: http://go/chrome-devtools:breakpoint-ux-design
Change-Id: Ia3e61882fc3a5d6563815f14644f1e54781456ae