[BreakpointsSidebarController] Remember expanded settings
This adds a setting to the sidebar controller to remember whether
certain groups were expanded/collapsed throughout different sessions.

It also auto-expands a group if a breakpoint within that group was hit,
or if a new breakpoint is added to that file.

Screenshot: https://imgur.com/a/F0uGmXz
Doc:  https://bit.ly/3pa4y92
Bug: 1324904, 1346231
Change-Id: If678162a914623db00cf004d410b580ce63d4c96