[BreakpointsSidebarPane] Add ability to edit breakpoints.
This adds a new button to the breakpoints view which shows up on hover,
right next to the remove button. When clicking the button, the sidebar
controller reveals the associated BreakpointLocation.

This also turns the BreakpointLocation into a proper class and adds a
revealer for it, implemented within the DebuggerPlugin, which upon
receiving a reveal event, will reveal the associated UILocation, and
will afterwards pop up the breakpoint edit dialog.

Bug: chromium:1346231
Fixed: chromium:1351752
Screenshot: https://imgur.com/XJuifYJ.png
Doc: http://go/chrome-devtools:breakpoint-ux-design
Change-Id: I454aaa102fce3884c5b139ce62ad49cddef3b5ed