[debugger] Speed up call stack function name lookup.
This is a follow-up fix for https://crrev.com/c/3720713, which
introduces the translation of function names in the Call Stack based on
source maps, and which introduced a performance regression that
specifically shows with huge Angular codebases.

The performance problem stems from the need to recompute line endings
for all scripts that appear in the (synchronous part of the) call stack
every time. The fix here is to generalize the caching of Text objects
that we already have in the NamesResolver to a global WeakMap, shared by
all the methods, which caches Text instances based on DeferredContent
instances (the results of ContentProvider.requestContent() calls). These
objects are stable for Script and UISourceCode, and are discarded upon
editing the content providers.

Fixed: chromium:1369002
Bug: chromium:1339683, chromium:424201
Change-Id: I20506b6b08d17f568d1bdb8b4471ab645c7aaa1f